screensetting vf, setting luminosity and tint to int according to kobo's driver, setting pagestep dynamically

+++ screensettings.h
@@ -25,8 +25,8 @@ private slots:
 
 private:
     Ui::ScreenSettings*  ui;
-    float                luminosity;
-    float                tint;
+    int                  luminosity;
+    int                  tint;
     bool                 settingAvailable;
     const KoboDeviceDescriptor& desc;
